requirements.txt
flask
flask_restful
flask_jwt_extended
werkzeug.security
flask_mail
flask_login
flask_sqlalchemy
flask_login
flask_mail
flask_cors
psycopg2
xlsx

@fortawesome/fontawesome-svg-core @fortawesome/free-solid-svg-icons @fortawesome/vue-fontawesome
# npm install xlsx